Universal treatment of blobs and lazy resolution of dependencies.
Added a blob onjectStore for storing binaries referenced by songs and stores
them with a compouned key of the song id and attribute name from which the blob
was created. The blobs are detatched from the song by setting the attribute to
false if no blob and true if one was stored.

This storage system is hidden behind the song model. When a binary is request,
the song provided inform the model, which retreives the blob, creates a data
url, and sets the attribut of the live song object to that url.

An attribut that references a binary will have one of three values:

  - False: No binary exists. Handle appropriately.

  - True: A binary exists, and can be resolved with the appropriate method in
    the song provider.

  - dataURL: The binary has already been resolved and the live data URL is the
    value of the attribute.